requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
fastapi==0.110.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
motor==3.7.1
openai==1.12.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.1.1
pycparser==3.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.16.0
pypdfium2==5.4.0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
regex==2024.5.15
sniffio==1.3.1
starlette==0.36.3
tqdm==4.67.3
typing_extensions==4.15.0
uvicorn==0.29.0
